Also give `AffineSpace` a `Generic` default definition.
This one is a bit less straightforward than `AdditiveGroup` etc.: we need
a new type representing the generic representation of the difference between
points in a combination of two spaces (product space). One option would be
to default to tuple, however the boxing would probably be inacceptable for
some applications, so introduce a new strict-tuple type for this purpose.